requirements.txt
# absl-py==1.4.0
# aiobotocore==2.12.1
# aiohttp==3.9.3
# aioitertools==0.11.0
# aiosignal==1.3.1
# anyio==3.7.1
# argon2-cffi==23.1.0
# argon2-cffi-bindings==21.2.0
arrow==1.3.0
asn1crypto==1.5.1
# asttokens==2.4.1
# astunparse==1.6.3
# async-lru==2.0.4
# async-timeout==4.0.3
# attrs==23.2.0
# Babel==2.14.0
# beautifulsoup4==4.12.3
# bleach==6.1.0
# botocore==1.34.51
cachetools==4.2.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudpickle==2.2.1
# colorama==0.4.6
# comm==0.2.1
# contourpy==1.2.0
# cryptography==42.0.5
# cycler==0.12.1
# debugpy==1.8.1
decorator==5.1.1
# defusedxml==0.7.1
# exceptiongroup==1.2.0
# executing==2.0.1
# fastjsonschema==2.19.1
# filelock==3.13.1
# flatbuffers==23.5.26
# fonttools==4.49.0
# fqdn==1.5.1
# frozenlist==1.4.1
# fsspec==2023.12.2
# gast==0.5.4
# google-auth==2.28.1
# google-auth-oauthlib==1.2.0
# google-pasta==0.2.0
grpcio==1.62.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
imbalanced-learn==0.10.1
imblearn==0.0
importlib-metadata==7.0.1
importlib-resources==5.13.0
# ipykernel==6.29.3
# ipython==8.18.1
# ipywidgets==8.1.2
# isoduration==20.11.0
# jedi==0.19.1
# Jinja2==3.1.3
# jmespath==1.0.1
joblib #@ file:///C:/b/abs_1anqjntpan/croot/joblib_1685113317150/work
json5==0.9.20
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
# jupyter==1.0.0
# jupyter-console==6.6.3
# jupyter-events==0.9.0
# jupyter-lsp==2.2.3
# jupyter_client==8.6.0
# jupyter_core==5.7.1
# jupyter_server==2.13.0
# jupyter_server_terminals==0.5.2
# jupyterlab==4.1.3
# jupyterlab_pygments==0.3.0
# jupyterlab_server==2.25.3
# jupyterlab_widgets==3.0.10
# keras==2.15.0
# kiwisolver==1.4.5
# libclang==16.0.6
# Markdown==3.5.2
# MarkupSafe==2.1.5
# matplotlib==3.7.1
# matplotlib-inline==0.1.6
# mistune==3.0.2
#mkl-fft #@ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
#mkl-random #@ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
#mkl-service==2.4.0
ml-dtypes==0.2.0
multidict==6.0.5
nbclient==0.9.0
nbconvert==7.16.2
nbformat==5.9.2
nest-asyncio==1.6.0
# notebook==7.1.1
# notebook_shim==0.2.4
numpy==1.23.5
oauthlib==3.2.2
# opt-einsum==3.3.0
# overrides==7.7.0
packaging==23.2
pandas==1.5.3
# pandocfilters==1.5.1
# parso==0.8.3
# pillow==10.2.0
# platformdirs==3.11.0
# prometheus_client==0.20.0
# prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
# pure-eval==0.2.2
pyarrow==15.0.0
# pyasn1==0.5.1
# pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
# pyOpenSSL==24.0.0
# pyparsing==3.1.1
# python-dateutil==2.9.0.post0
# python-json-logger==2.0.7
# pytimeparse==1.1.8
# pytorch-tabnet==3.1.1
# pytz==2024.1
#pywin32==306
# pywinpty==2.0.13
# PyYAML==6.0.1
# pyzmq==25.1.2
# qtconsole==5.5.1
# QtPy==2.4.1
# referencing==0.33.0
# requests==2.31.0
# requests-oauthlib==1.3.1
# retrying==1.3.4
# rfc3339-validator==0.1.4
# rfc3986-validator==0.1.1
# rpds-py==0.18.0
# rsa==4.9
# s3fs==2023.12.2
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
# Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.1
snowflake-connector-python==3.7.1
snowflake-ml-python==1.2.3
snowflake-snowpark-python==1.13.0
snowflake.core==0.8.0
sortedcontainers==2.4.0
#soupsieve==2.5
sqlparse==0.4.4
#stack-data==0.6.3
# tensorboard==2.15.2
# tensorboard-data-server==0.7.2
# tensorflow==2.15.0
# tensorflow-estimator==2.15.0
# tensorflow-intel==2.15.0
# tensorflow-io-gcs-filesystem==0.31.0
# termcolor==2.4.0
# terminado==0.18.0
# threadpoolctl #@ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
# tinycss2==1.2.1
# tomli==2.0.1
# tomlkit==0.12.4
# #torch==1.10.2
# tornado==6.4
# tqdm==4.66.2
# traitlets==5.14.1
# types-python-dateutil==2.8.19.20240106
# typing_extensions==4.10.0
# uri-template==1.3.0
# urllib3==1.26.18
# wcwidth==0.2.13
# webcolors==1.13
# webencodings==0.5.1
# websocket-client==1.7.0
# Werkzeug==3.0.1
# widgetsnbextension==4.0.10
#wrapt==1.14.1
xgboost==1.7.3
#yarl==1.9.4
zipp==3.17.0
